Move tryCompiletimeConstantFold to separate file
